Added ENT:SetSize method
Added ENT:SetSize in order to prepare the entity as a base. This method is available on both server and client, but it's preferred to be used on the server only.
Added OnScaledBoxSizeChange hook, called any time an entity changes its size to help external addons update their physical object settings.